add log to provide hint in debugging if serverCertificate was unintionally omitted.
When using a keySystem that needs a serverCertificate, if by mistake user forgot to set the `serverCertificate` options,
the CDM is still able to create a challenge but such a challenge will be too short (2 bytes on widevine) and the license request will be refused by the license server upon received that unsuitable challenge.

Usually theses challenge are much longer.
This commit add a log warning if the challenged is considered too short to be suitable.
This log warning intent is to help debugging someone that has unintionally omitted to set the `serverCertficate` options.