[Proposal] On the PlayStation 4, update the tfdt instead of using timestampOffset
The issue
---------

We recently noticed that the PlayStation 4's browser had issues
with the MSE `timestampOffset` `SourceBuffer` attribute, which
allows to add an offset to pushed media segments relative to their
"regular" starting timestamp.

More specifically, it seemed to properly append the media data at the right
timestamp when interrogating both the `SourceBuffer`s and the
`HTMLMediaElement`'s `buffered` property, but playback was not able to
start.

To work-around it, we noticed that updating instead the starting
timestamp at the container level (e.g. on the pushed mp4 files
directly) did have the expected result.

This proposal
-------------

Here I propose to add compatibility code at the level of the
`SourceBuffer.appendBuffer` MSE API, which allows to push segments.

Here, a new `appendBufferWithOffset` function would be able to check,
based on device-detection (usually through the user agent), whether the
timestamp should be added by relying on the `timestampOffset` attribute
(on all devices except the PlayStation 4, and even on the PlayStation 4
if the container file is not recognized, see below) or by patching the
container file (only on the PlayStation 4 on recognized container
formats).

This is not as straightforward as it first may seem, because to patch
mp4 media segments, we need to rely on a timescale value (as the
corresponding `trackFragmentDecodeTime` is always stored as an integer
in ISOBMFF) which is announced in the initialization segment - meaning
there's the need for some state. It could be done in several ways
(a class, a value stored by the caller etc.), but for now I opted for
a WeakMap-based solution (which maps each `SourceBuffer` to its
corresponding timescale).

This solution isn't the simplest one to understand as is but allows to
just isolate the complexity to the compat files, which I judged more
readable.

Limitations
-----------

There are several limitations with the current implementation:

  - Theoretically, it is possible in MSE to tell the browser to
    basically "forget" about the last initialization segment pushed,
    through the `SourceBuffer.abort` method.

    If we wanted to be thorough, we should probably also remove all
    memory about the last init segment's timescale stored on that
    SourceBuffer.
    I However didn't do that because in all real scenarios I could think
    of, there was none where this could be an issue (famous last words).

  - For now in that proposal, I only parse ISOBMFF files (basically MP4
    files).

    We may be able to do the same thing for webm/mkv files with few
    effort, to check.

    For all scenarios where we don't recognize the container, we
    fallback to using the MSE `SourceBuffer.timestampOffset` attribute
    as a last resort.

  - There may be rounding error at some point, or even when factoring in
    the timescale. In observed scenarios, this did not lead to
    problematic scenarios on devices for now.

  - It adds a lot of code.

    Note that we also have a SourceBuffer patch specific to old webkit
    devices (among which maybe the PlayStation 4 :D, I did not check
    that but it would be kind of funny), that actually monkey-patches
    the `WebkitSourceBuffer` native class (this is the only ugly
    instance of a monkey patch in the RxPlayer codebase).

    I though about merging the two into a single cleaner work-around but
    I was both too lazy for now and afraid to break things for only a
    proposal, so I did not do so yet.