Fix memory tests on Chrome 124
Our memory tests (which checks for memory leaks) are failing since Chrome 124.

It seems that the issue is not memory leaks on our part but changes in
the way Chrome performs its garbage collection, especially after the
`gc` global function (it exposes it only if the `--js-flags="--expose-gc"`
chrome flag is used when running) call.

Awaiting several seconds after that call seems to improve measures by a
lot (tens of MB being collected), though not every tests seems to pass
yet, so that may not be the only change we'll need to do.