Fix integration tests
It seems that some integration tests often fail in the CI.

The reasons are mostly unimportant: e.g. most often those are "Uncaught
Error" provoked due to a race condition between the logic running in the
worker performs an operation just as the test finished and the content
is stopped. I'm thinking here especially with the `endOfStream` API when
MSE is available in a Worker environment (so, on Chrome).

Though this error should lead to no real issue, having uncaught error
could lead to unnecessary reporting by e.g. browser inspectors and such
(and... integration tests), so I just transform the one spotted into an
error log.