Remove the need for the express dependency
We were relying on express only for serving static files for development
builds and testing purposes - in a way that was kind of overkill considering
that we need almost none of its features.

To simplify maintainance, I decided to remove that dependency and just
implement the bare minimum we need from it instead.