Remove now unneeded dependencies (webpack, babel, terser)
This commit removes dependencies that should now be unneeded.

I had to update the standalone demo's script to remove its dependency on
webpack by copy-pasting most of its bundling code from the full demo's
script.

As we never really had a use for the standalone demo, I'll also propose
in the future to just remove it, but I prefer to do that in a second
step.