CMCD: Add nrr and nor handling
Previous iteration of CMCD [#1461] did not handle "nrr" nor... "nor"
attributes.

The former (nrr) gives a hint to the CDN of what the next byte range
request going to be (if it's one) and the later (nor) indicates the
relative URL based on the current one.

Note that if it's the same segment, the relative URL is just the file
name, as it seems logical for me - and as I didn't find an explicit
mention of this frequent case in the CMCD specification.
Even if the documentation for `nrr` indicates that an absent `nor`
but a present `nrr` is assumed to mean the same segment.