[Proposal] add `checkManifestIntegrity` option
Years ago, after noticing a rare issue of media segment corruption
either on the CDN or packaging-side, we realized that the back-end
cannot always be trusted to deliver a full well-formed response, even on
an HTTP(S) 200.

After initially waiting for an investigation and fix, during which
some customers would have a fatal playback error when the issue happened,
we finally decided to add the `checkMediaSegmentIntegrity` resilience
option.

That option just checks, when an ISOBMFF segment is requested, that the
top-level boxes aren't truncated (and if they are, we treat the request
as we would have treated a retry-able request error).

---

We've now proof that the same thing also very rarely happen for Manifest
requests, especially at times where a very large audience of people are
watching a live stream on some CDN.

Likewise, we're waiting for some investigation and fix and in the
meantime, customers are seeing errors.

I here reluctantly propose the `checkManifestIntegrity` option, which
will do very minimal checks to see if the Manifest is truncated after
its request - and if it is, we retry that request through our usual
exponential backoff rules.

I only do so for DASH MPD for now (the option has no effect for other
transport) and only check that it ends with `\</MPD\s>\s` (with `\s`
meaning XML white space).
I do that minimal check instead of just checking for a corrupted data in
the XML because doing the latter is both much harder to implement (full
parsing is done by our parser which is not part of the request-retry
algorithm) and is going against our performance-first strategy on MPD
parsing.

Note that because the check is minimal, it does not catch ALL seen case
of Manifest corruptions, but it does for some.