Remove fake server in tests and the necessity of using sinonjs
We were still relying on a complex XHR patch exposed by sinonjs in our
integration tests - the last part of our codebase that needed sinon - to
test our manifest retry logic.

This was unneeded as we can just rely on our `manifestLoader` API and
return a retry-able error with the same effect, without the need of any
complex and incomplete monkey-patching.

The potential for this commit was seen while PoCing new test strategies where
both requests and MSE+EME API are completely mocked (to facilitate adding
new tests, as that would mean we wouldn't have to provide real media
segments and licenses anymore for each scenario).
After first trying to find a maintainable way to patch our request
logic, I ended up founding that relying on our `manifestLoader` and
`segmentLoader` API may in fact be much simpler and is already available.