Add `checkManifestIntegrity` and `checkMediaSegmentIntegrity` checkboxes to the demo
In #1471, I added the `checkManifestIntegrity` option because we noticed
issues at Canal+ of incomplete MPD delivered by the CDN.

This is most probably a back-end issue, but we added that option because of
the slow progression in fixing that issue.

Here, I propose to add both that check and the older
`checkMediaSegmentIntegrity` check to the demo page. This might help us
debugging issues in the future yet is also useful to showcase those options
to people looking at our player first without diving into our API
documentation.

Note that the `Request Configuration` part of our settings becomes quite
large but this isn't much of an issue for me.