Remove the need for the DOMParser when loading a smooth content
This is a continuation of our work on DASH toward removing reliance on
the DOMParser in profit of our own DOM parser - generally faster and
compatible to more environments (WebWorkers, Nodejs...).

It could also theoretically mean that we'll be able to run Smooth in the
worker even if we do not have an API for adding features worker-side
yet, nor really the need for it yet.