[POC] Update `Manifest` structure to make HLS implementation possible
NOTE: This is just a background, low-priority, Proof-Of-Concept.
It is not currently functional and is in the middle of its
implementation, there's also no HLS code for now, only HLS-related
ideas.

Overview
========

This PR is a Proof-Of-Concept where I explore whether HLS playback can
be natively implemented in the RxPlayer without sacrificing too much the
RxPlayer code's readability. This is not the first attempt, but the
approach here is different than previous ones.

One of the core idea is to keep the RxPlayer API as is, and thus to hide
all HLS differences in the RxPlayer code.

We already thought of potential HLS compatibility when designing our v4, so
there seem to be nothing in our API incompatible with HLS concepts.
Also our core code is for the most part protocol-agnostic, so we presumably
could keep, and profit from, most of our logic while playing HLS streams.

However, this attempt completely changes our internal `Manifest` concept,
especially the `Adaptation` subpart of it (which was equivalent to the
concept of a "track").

Manifest hierarchy change
=========================

The issue
---------

One of the main differences between HLS and other protocols handled by the
RxPlayer (such as DASH) is the concept of "variant streams" only HLS have.

Basically in DASH, you first select your wanted audio + video + text tracks
and then select a Representation (i.e. quality) according to your current
playback conditions such as your network bandwidth.

In HLS however, this is reversed. You first have to select your "variant
stream" based on your playback conditions (mainly: network bandwidth) and
then see which "media" - including our notion of a track - is available for
that variant stream.

Because of edge cases, we cannot translate the HLS model into our DASH-like
model wihout losing some HLS features.
Doing the reverse (adapting DASH model into HLS model, which seems to be
what e.g. the shaka-player more-or-less did) could work yet it doesn't seem
like an optimal solution (e.g. we would either lose some bandwidth
information or creating a huge number of variant streams for all potential
combinations).

How I tried to fix it
---------------------

So what I did here, was to change the Manifest object's hierarchy so it
can adapt to both how HLS is treating thing and how DASH is treating
things.

Previously, it could be summarized as such:
```
Manifest -> Period -> Adaptation (i.e. track) -> Representation (i.e. quality)
```
Basically Manifest contains Periods (time-delimited subparts of a contents),
each with its own audio, video and text Adaptation (tracks), each with their
own Representations (qualities).

In this Proof-Of-Concept, I'm reworking it into:
```
Manifest -> Period -> VariantStream -> Media -> Representation
```

Where:

**VariantStream**: This is a grouping of tracks that should be considered
under specific conditions (e.g. a network bandwidth):

  - For HLS, this is equivalent to the "variant stream" concept.

  - For DASH, HSS and others, there's no such concept (the network bandwidth
    for example is relied on at the Representation-level instead).
    Thus they posses a single VariantStream, with no condition, containing
    all tracks available for that period under the "Media" concept.

**Media**: subset of an audio, video or text track which is part of a
VariantStream (a "track" may be scattered into multiple variant streams,
in which case we would have multiple "media" in several VariantStreams
linked to the same track).

  - For HLS: those are the media associated to a variant.
    Note that there should only be a single Representation per media here,
    as HLS already differentiate between bandwidth at the variant stream
    level.

  - For DASH, HSS and others: as they only have a single VariantStream, a
    media is equivalent to a track and contains directly contains all the
    Representations for that track.

Adaptation / tracks in all that
-------------------------------

An issue with that solution is that we lose the concept of an
"Adaptation", that was nicely a one-to-one translation to the API concept
of a "track".

Here, "tracks" conceptually still exist (e.g. there could be in HLS a
french audio "media" available in most variant streams), yet they may be
scattered as multiple "media" in various "VariantStream".

To facilitate the handling of "track" in the API and core code, I also
added in a `Period` object the `tracksMetadata` property, which just
summarize the characteristics of all tracks which may be found in the various
VariantStreams of that `Period`.

To avoid repeating properties in both `tracksMetadata` and in the
VariantStreams' "Media", the latter then refers to which track it is
linked to using a `linkedTrackId` property (so just referring to the track
through an id):
```ts
interface IPeriodMetadata {
 id: string;
 start: number;
 end?: number | undefined;
 duration?: number | undefined;
 streamEvents: IManifestStreamEvent[];
 tracksMetadata: {
   // Links audio track ids to their metadata
   audio: Record<string, IAudioTrackMetadata>;
   video: Record<string, IVideoTrackMetadata>;
   text: Record<string, ITextTrackMetadata>;
 };
 variantStreams: Array<{
   id: string;
   bandwidth: number | undefined;
   audio: Array<{
     id: string;
     // Corresponding audio track id
     linkedTrackId: string;
     representations: IRepresentationMetadata[];
   }>;
   // same for video and text...
 }>;
}
```